[KOGITO-5718] Fix bogus Quarkus WARN for namespace clash
