+++ kogito-codegen-modules/kogito-codegen-processes/src/main/resources/class-templates/EventReceiverQuarkusTemplate.java
@@ -23,7 +23,7 @@ import javax.enterprise.context.ApplicationScoped;
 
 import org.eclipse.microprofile.reactive.messaging.Incoming;
 import org.eclipse.microprofile.reactive.messaging.Message;
-import org.kie.kogito.addon.cloudevents.quarkus.AbstractQuarkusCloudEventReceiver;
+import org.kie.kogito.addon.quarkus.messaging.common.AbstractQuarkusCloudEventReceiver ;
 
 import io.quarkus.runtime.Startup;
