+++ quarkus/addons/messaging/common/src/main/java/org/kie/kogito/addon/quarkus/messaging/common/message/MessageDecorator.java
@@ -13,13 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.kogito.app;
+package org.kie.kogito.addon.quarkus.messaging.common.message;
 
-import javax.ws.rs.Path;
+import org.eclipse.microprofile.reactive.messaging.Message;
 
-import org.kie.kogito.addon.quarkus.messaging.common.http.AbstractQuarkusCloudEventResource;
+/**
+ * {@link MessageDecorator}s can decorates the {@link Message} envelope with metadata and additional information in a given context.
+ */
+public interface MessageDecorator {
 
-@Path("/")
-public class CloudEventListenerResource extends AbstractQuarkusCloudEventResource {
-
+    /**
+     * Decorates the given reactive message
+     *
+     * @param payload payload to decorate
+     * @param <T> payload type
+     * @return payload in Message format decorated
+     */
+    <T> Message<T> decorate(T payload);
 }
