+++ quarkus/addons/messaging/integration-tests/src/test/java/org/kie/kogito/integrationtests/QuarkusKogitoExtensionInitializerIT.java
@@ -47,7 +47,7 @@ public class QuarkusKogitoExtensionInitializerIT {
         KogitoExtension kogitoExtension = ExtensionProvider.getInstance().parseExtension(KogitoExtension.class, event);
 
         assertNotNull(kogitoExtension, "KogitoExtension not registered, please make sure " +
-                "bean org.kie.kogito.addon.cloudevents.quarkus.QuarkusKogitoExtensionInitializer has been loaded");
+                "bean org.kie.kogito.addon.quarkus.messaging.common.QuarkusKogitoExtensionInitializer has been loaded");
         assertEquals(DecisionTestUtils.MODEL_NAME, kogitoExtension.getDmnModelName());
         assertEquals(DecisionTestUtils.MODEL_NAMESPACE, kogitoExtension.getDmnModelNamespace());
     }
